intento2 de eleminar matriz

+++ Hash.h
@@ -1,6 +1,6 @@
 #include "State.h"
 struct Node {
-    long long unsigned int state_id;
+    long int state_id;
     Node* next;
     Node(int state_id){
         this->state_id = state_id;
@@ -11,12 +11,12 @@ typedef struct Node Node;
 class Hash {
 private:
     Node* hashTable[1000];
-    int hashFunction(long long unsigned int state);
+    int hashFunction(long int state);
 public:
     Hash();
     ~Hash();
-    void push(long long unsigned int state);
-    bool find(long long unsigned int state);
+    void push(long int state);
+    bool find(long int state);
     void clear();
     void print();
 };
